Combine exporter and cacher phases
Combines the exporter and cacher phases to reuse layer tarballs.

* Move logic from cacher.go to exporter.go.
* Move tests from cacher_test.go to exporter_test.go.
* Rename cache image flag from '-image' to '-cache-image'.
* Rename cache directory flag from '-path' to '-cache-dir'.
* Remove unused 'In' field of the Exporter struct.
* Remove cacher implementation, tests, and command.
