+++ exporter.go
@@ -3,6 +3,7 @@ package lifecycle
 import (
 	"encoding/json"
 	"fmt"
+	"io"
 	"path/filepath"
 
 	"github.com/BurntSushi/toml"
@@ -13,12 +14,23 @@ import (
 	"github.com/buildpack/lifecycle/cmd"
 )
 
+//go:generate mockgen -package testmock -destination testmock/cache.go github.com/buildpack/lifecycle Cache
+type Cache interface {
+	Name() string
+	SetMetadata(metadata CacheMetadata) error
+	RetrieveMetadata() (CacheMetadata, error)
+	AddLayerFile(sha string, tarPath string) error
+	ReuseLayer(sha string) error
+	RetrieveLayer(sha string) (io.ReadCloser, error)
+	Commit() error
+}
+
 type Exporter struct {
 	Buildpacks   []Buildpack
 	ArtifactsDir string
-	In           []byte
 	Logger       Logger
 	UID, GID     int
+	tarHashes    map[string]string // Stores hashes of layer tarballs for reuse between the export and cache steps.
 }
 
 type LauncherConfig struct {
@@ -39,26 +51,26 @@ func (e *Exporter) Export(
 	var err error
 
 	meta := LayersMetadata{}
-
 	meta.RunImage.TopLayer, err = workingImage.TopLayer()
 	if err != nil {
 		return errors.Wrap(err, "get run image top layer SHA")
 	}
 
+	// TODO why are we saving stack in addition to runImageRef when the two can be different?
 	meta.RunImage.Reference = runImageRef
 	meta.Stack = stack
 
-	meta.App.SHA, err = e.addLayer(workingImage, &layer{path: appDir, identifier: "app"}, origMetadata.App.SHA)
+	meta.App.SHA, err = e.addOrReuseLayer(workingImage, &layer{path: appDir, identifier: "app"}, origMetadata.App.SHA)
 	if err != nil {
 		return errors.Wrap(err, "exporting app layer")
 	}
 
-	meta.Config.SHA, err = e.addLayer(workingImage, &layer{path: filepath.Join(layersDir, "config"), identifier: "config"}, origMetadata.Config.SHA)
+	meta.Config.SHA, err = e.addOrReuseLayer(workingImage, &layer{path: filepath.Join(layersDir, "config"), identifier: "config"}, origMetadata.Config.SHA)
 	if err != nil {
 		return errors.Wrap(err, "exporting config layer")
 	}
 
-	meta.Launcher.SHA, err = e.addLayer(workingImage, &layer{path: launcherConfig.Path, identifier: "launcher"}, origMetadata.Launcher.SHA)
+	meta.Launcher.SHA, err = e.addOrReuseLayer(workingImage, &layer{path: launcherConfig.Path, identifier: "launcher"}, origMetadata.Launcher.SHA)
 	if err != nil {
 		return errors.Wrap(err, "exporting launcher layer")
 	}
@@ -68,10 +80,13 @@ func (e *Exporter) Export(
 		if err != nil {
 			return errors.Wrapf(err, "reading layers for buildpack '%s'", bp.ID)
 		}
-		bpMD := BuildpackLayersMetadata{ID: bp.ID, Version: bp.Version, Layers: map[string]BuildpackLayerMetadata{}}
-
-		layers := bpDir.findLayers(launch)
-		for i, layer := range layers {
+		bpMD := BuildpackLayersMetadata{
+			ID:      bp.ID,
+			Version: bp.Version,
+			Layers:  map[string]BuildpackLayerMetadata{},
+		}
+		for _, layer := range bpDir.findLayers(launch) {
+			layer := layer
 			lmd, err := layer.read()
 			if err != nil {
 				return errors.Wrapf(err, "reading '%s' metadata", layer.Identifier())
@@ -79,7 +94,7 @@ func (e *Exporter) Export(
 
 			if layer.hasLocalContents() {
 				origLayerMetadata := origMetadata.MetadataForBuildpack(bp.ID).Layers[layer.name()]
-				lmd.SHA, err = e.addLayer(workingImage, &layers[i], origLayerMetadata.SHA)
+				lmd.SHA, err = e.addOrReuseLayer(workingImage, &layer, origLayerMetadata.SHA)
 				if err != nil {
 					return err
 				}
@@ -101,6 +116,7 @@ func (e *Exporter) Export(
 			}
 			bpMD.Layers[layer.name()] = lmd
 		}
+		meta.Buildpacks = append(meta.Buildpacks, bpMD)
 
 		if malformedLayers := bpDir.findLayers(malformed); len(malformedLayers) > 0 {
 			ids := make([]string, 0, len(malformedLayers))
@@ -109,8 +125,6 @@ func (e *Exporter) Export(
 			}
 			return fmt.Errorf("failed to parse metadata for layers '%s'", ids)
 		}
-
-		meta.Buildpacks = append(meta.Buildpacks, bpMD)
 	}
 
 	data, err := json.Marshal(meta)
@@ -154,11 +168,75 @@ func (e *Exporter) Export(
 	return saveImage(workingImage, additionalNames, e.Logger)
 }
 
-func (e *Exporter) addLayer(image imgutil.Image, layer identifiableLayer, previousSHA string) (string, error) {
+func (e *Exporter) Cache(layersDir string, cacheStore Cache) error {
+	var err error
+	origMeta, err := cacheStore.RetrieveMetadata()
+	if err != nil {
+		return errors.Wrap(err, "metadata for previous cache")
+	}
+	meta := CacheMetadata{}
+
+	for _, bp := range e.Buildpacks {
+		bpDir, err := readBuildpackLayersDir(layersDir, bp)
+		if err != nil {
+			return errors.Wrapf(err, "reading layers for buildpack '%s'", bp.ID)
+		}
+
+		bpMD := BuildpackLayersMetadata{
+			ID:      bp.ID,
+			Version: bp.Version,
+			Layers:  map[string]BuildpackLayerMetadata{},
+		}
+		for _, layer := range bpDir.findLayers(cached) {
+			layer := layer
+			if !layer.hasLocalContents() {
+				return fmt.Errorf("failed to cache layer '%s' because it has no contents", layer.Identifier())
+			}
+			lmd, err := layer.read()
+			if err != nil {
+				return errors.Wrapf(err, "reading %q metadata", layer.Identifier())
+			}
+			origLayerMetadata := origMeta.MetadataForBuildpack(bp.ID).Layers[layer.name()]
+			if lmd.SHA, err = e.addOrReuseCacheLayer(cacheStore, &layer, origLayerMetadata.SHA); err != nil {
+				return err
+			}
+			bpMD.Layers[layer.name()] = lmd
+		}
+		meta.Buildpacks = append(meta.Buildpacks, bpMD)
+	}
+
+	if err := cacheStore.SetMetadata(meta); err != nil {
+		return errors.Wrap(err, "setting cache metadata")
+	}
+	if err := cacheStore.Commit(); err != nil {
+		return errors.Wrap(err, "committing cache")
+	}
+
+	return nil
+}
+
+func (e *Exporter) tarLayer(layer identifiableLayer) (string, string, error) {
 	tarPath := filepath.Join(e.ArtifactsDir, escapeID(layer.Identifier())+".tar")
+	if e.tarHashes == nil {
+		e.tarHashes = make(map[string]string)
+	}
+	if sha, ok := e.tarHashes[tarPath]; ok {
+		e.Logger.Debugf("Reusing tarball for layer %q with SHA: %s\n", layer.Identifier(), sha)
+		return tarPath, sha, nil
+	}
+	e.Logger.Debugf("Writing tarball for layer %q\n", layer.Identifier())
 	sha, err := archive.WriteTarFile(layer.Path(), tarPath, e.UID, e.GID)
 	if err != nil {
-		return "", errors.Wrapf(err, "exporting layer '%s'", layer.Identifier())
+		return "", "", err
+	}
+	e.tarHashes[tarPath] = sha
+	return tarPath, sha, nil
+}
+
+func (e *Exporter) addOrReuseLayer(image imgutil.Image, layer identifiableLayer, previousSHA string) (string, error) {
+	tarPath, sha, err := e.tarLayer(layer)
+	if err != nil {
+		return "", errors.Wrapf(err, "tarring layer '%s'", layer.Identifier())
 	}
 	if sha == previousSHA {
 		e.Logger.Infof("Reusing layer '%s'\n", layer.Identifier())
@@ -168,4 +246,19 @@ func (e *Exporter) addLayer(image imgutil.Image, layer identifiableLayer, previo
 	e.Logger.Infof("Adding layer '%s'\n", layer.Identifier())
 	e.Logger.Debugf("Layer '%s' SHA: %s\n", layer.Identifier(), sha)
 	return sha, image.AddLayer(tarPath)
+}
+
+func (e *Exporter) addOrReuseCacheLayer(cache Cache, layer identifiableLayer, previousSHA string) (string, error) {
+	tarPath, sha, err := e.tarLayer(layer)
+	if err != nil {
+		return "", errors.Wrapf(err, "tarring layer %q", layer.Identifier())
+	}
+	if sha == previousSHA {
+		e.Logger.Infof("Reusing cache layer '%s'\n", layer.Identifier())
+		e.Logger.Debugf("Layer '%s' SHA: %s\n", layer.Identifier(), sha)
+		return sha, cache.ReuseLayer(previousSHA)
+	}
+	e.Logger.Infof("Adding cache layer '%s'\n", layer.Identifier())
+	e.Logger.Debugf("Layer '%s' SHA: %s\n", layer.Identifier(), sha)
+	return sha, cache.AddLayerFile(sha, tarPath)
 }
